packages/forge-train/pyproject.toml
[project]
name = "forge-train"
version = "0.1.0"
description = "Training pipeline for diffusion models: LoRA and DoRA fine-tuning"
requires-python = ">=3.11,<3.13"
dependencies = [
    "forge-core",
    "diffusers>=0.31",
    "transformers>=4.44",
    "accelerate>=1.0",
    "peft>=0.14",
    "safetensors>=0.4.5",
    "torch>=2.5,<2.8",
    "torchvision>=0.20",
    "pillow>=11.0",
    "numpy>=1.26",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
cuda = [
    "bitsandbytes>=0.44",
    "xformers>=0.0.28",
]

[project.scripts]
forge-train = "forge.train.pipeline:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/forge"]

[tool.uv.sources]
forge-core = { workspace = true }
